+++ HISTORY.md
@@ -5,6 +5,7 @@
 * Added pickle load inspection [PIC100](doc/checks/PIC100.md)
 * Added django safe strings inspection [DJG102](doc/checks/DJG102.md)
 * Added hardcoded temp path read or write inspection [TMP101](doc/checks/TMP101.md)
+* Added XML standard library DoS inspection [XML100](doc/checks/XML100.md)
 * Added XML RPC dotted paths inspection [XML200](doc/checks/XML200.md)
 
 ## 1.6.0